+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+
 # Changelog
 
 All notable changes to this project will be documented in this file.
@@ -15,6 +16,17 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 ## [UNRELEASED]
 
+### Added
+* Pretty print the `rpu_config` in a readable manner (excluding all the default
+  settings). (\#60)
+* `ReferenceUnitCell`  is introduced, which has two devices, where one is fixed and
+  the other updated and the effective weight is computed a difference
+  betweem the two. (\#61)
+* `VectorUnitCell` accepts now arbitrariy weighting schemes that can
+  be user-defined by using a new `gamma_vec` property that spcified
+  how to combine the unit cell devices to form the
+  effective weight. (\#61)
+
 ### Changed
 
 * The unit cell items in `aihwkit.simulator.configs` have been renamed, removing
@@ -24,8 +36,8 @@ The format is based on [Keep a Changelog], and this project adheres to
 * The `pybind11` version required has been bumped to 2.6.0, which can be
   installed from `pip` and makes system-wide installation no longer required.
   Please update your `pybind11` accordingly for compiling the library. (\#44)
-* Pretty print the `rpu_config` in a readable manner (excluding all the default 
-  settings). (\#60)
+* `VectorUnitCell` properties renamed and extended with an update policy specifying
+  how to select the hidden devices. (#61)
 
 ### Removed